Networks and Sites: Revert the use of the metadata API for `*_network_options` functions.
[54080] refactored the logic in `get_network_option()`, `update_network_option()` and `delete_network_option()` to use the metadata API. However, this change resulted in issues with large multisite installs that utilize memcached having network options > 1MB in size.

This change reverts [54080] and all related follow-up changes.

Reverts [54080], [54081], and [54082].  Partially reverts [54267] and [54402].

Props pavelschoffer, rebasaurus, johnbillion, spacedmonkey, desrosj, rinatkhaziev.
Fixes #56845.
See #37181.
Built from https://develop.svn.wordpress.org/trunk@54637


git-svn-id: http://core.svn.wordpress.org/trunk@54189 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-includes/ms-functions.php
@@ -113,7 +113,7 @@ function get_active_blog_for_user( $user_id ) {
  * @return int Number of active sites on the network.
  */
 function get_blog_count( $network_id = null ) {
-	return (int) get_network_option( $network_id, 'blog_count' );
+	return get_network_option( $network_id, 'blog_count' );
 }
 
 /**
@@ -2563,7 +2563,7 @@ function get_space_allowed() {
 	 *
 	 * @param int $space_allowed Upload quota in megabytes for the current blog.
 	 */
-	return (int) apply_filters( 'get_space_allowed', $space_allowed );
+	return apply_filters( 'get_space_allowed', $space_allowed );
 }
 
 /**
